Conflicts:
	aura/src/main/webapp/WEB-INF/views/reviewPost.jsp

+++ aura/src/main/java/com/bitcamp/aura/comment/model/CommentVO.java
@@ -20,7 +20,7 @@ public class CommentVO {
     private String nickname;
     private String comment_Contents;
     private int comment_Like;
-    private int comment_Score;
+    private float comment_Score;
     private String comment_Date;
     private int comment_Num;
 }
